pyproject.toml
[tool.poetry]
name = "ttp101"
version = "0.1.2"
description = ""
authors = ["ubaumann <github@m.ubaumann.ch>"]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3"
pyaml = "^23"
gunicorn = "^21"
ttp = "^0.9"

[tool.poetry.dev-dependencies]
black = "^23"
pylama = "^8"
rich = "^13"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
